refactor: separates out ai class
